aiagency/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
redis==5.1.1
sse-starlette==2.1.3
openai==1.51.0
anthropic==0.34.2
